+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `diffusion` module:
     - Function `euler_explicit1D`: Computes diffusion on a 1D space over
       a time-step using Euler explicit method.
+- Renamed `databases.up_query` to `databases.up_map`
 
 ## v0.0.5:
 - Added `databases` module: